chore(viewer): move HTML sanitizing to own component
