+++ packages/form-js-viewer/src/render/components/Util.js
@@ -1,5 +1,7 @@
 import snarkdown from 'snarkdown';
 
+import { sanitizeHTML } from './Sanitizer';
+
 export function formFieldClasses(type, errors = []) {
   if (!type) {
     throw new Error('type required');
@@ -21,28 +23,6 @@ export function prefixId(id) {
   return `fjs-form-${ id }`;
 }
 
-const NODE_TYPE_TEXT = 3,
-      NODE_TYPE_ELEMENT = 1;
-
-const DISALLOWED_NODES = [
-  'embed',
-  'iframe',
-  'object',
-  'script',
-  'svg'
-];
-
-const ALLOWED_ATTRIBUTES = [
-  'align',
-  'alt',
-  'class',
-  'href',
-  'id',
-  'name',
-  'src',
-  'valign'
-];
-
 export function markdownToHTML(markdown) {
   const htmls = markdown
     .split(/(?:\r?\n){2,}/)
@@ -59,44 +39,5 @@ export function markdownToHTML(markdown) {
 export function safeMarkdown(markdown) {
   const html = markdownToHTML(markdown);
 
-  const doc = new DOMParser().parseFromString(
-    `<!DOCTYPE html>\n<html><body><div>${ html }`,
-    'text/html'
-  );
-
-  doc.normalize();
-
-  sanitize(doc.body);
-
-  return new XMLSerializer().serializeToString(doc.body.firstChild);
-}
-
-function sanitize(node) {
-  if (node.nodeType === NODE_TYPE_TEXT) {
-    return;
-  }
-
-  if (node.nodeType !== NODE_TYPE_ELEMENT || DISALLOWED_NODES.includes(node.tagName.toLowerCase())) {
-    return node.remove();
-  }
-
-  for (let i = node.attributes.length; i--;) {
-    const name = node.attributes[ i ].name;
-
-    if (!ALLOWED_ATTRIBUTES.includes(name.toLowerCase())) {
-      node.attributes.removeNamedItem(name);
-    }
-
-    if (name === 'href') {
-      const href = node.attributes.getNamedItem('href');
-
-      if (href.value.replace(/\n/g, '').includes('javascript:')) {
-        node.attributes.removeNamedItem('href');
-      }
-    }
-  }
-
-  for (let i = node.childNodes.length; i--;) {
-    sanitize(node.childNodes[ i ]);
-  }
+  return sanitizeHTML(html);
 }
